Split record-caching into two dedicated endpoints, update api docs and requests
